Move internal classes to unexporter package
