+++ choreo-extension-native/spotbugs-exclude.xml
@@ -25,11 +25,11 @@
         <Bug pattern="RV_RETURN_VALUE_IGNORED_BAD_PRACTICE,NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE"/>
     </Match>
     <Match>
-        <Class name="io.ballerina.observe.choreo.client.secret.AnonymousAppSecretHandler" />
+        <Class name="io.ballerina.observe.choreo.client.internal.secret.AnonymousAppSecretHandler" />
         <Bug pattern="RV_RETURN_VALUE_IGNORED_BAD_PRACTICE,NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE"/>
     </Match>
     <Match>
-        <Class name="io.ballerina.observe.choreo.client.secret.AnonymousAppSecretHandler"/>
+        <Class name="io.ballerina.observe.choreo.client.internal.secret.AnonymousAppSecretHandler"/>
     </Match>
     <Match>
         <Class name="io.ballerina.observe.choreo.client.ChoreoSpanExporter"/>
