+++ choreo-extension-native/src/main/java/io/ballerina/observe/choreo/client/ChoreoClient.java
@@ -20,6 +20,9 @@ package io.ballerina.observe.choreo.client;
 
 import io.ballerina.observe.choreo.client.error.ChoreoClientException;
 import io.ballerina.observe.choreo.client.error.ChoreoErrors;
+import io.ballerina.observe.choreo.client.model.ChoreoMetric;
+import io.ballerina.observe.choreo.client.model.ChoreoTraceSpan;
+import io.ballerina.observe.choreo.client.model.SpanEvent;
 import io.ballerina.observe.choreo.gen.HandshakeGrpc;
 import io.ballerina.observe.choreo.gen.HandshakeOuterClass;
 import io.ballerina.observe.choreo.gen.HandshakeOuterClass.PublishAstRequest;
@@ -28,9 +31,6 @@ import io.ballerina.observe.choreo.gen.TelemetryGrpc;
 import io.ballerina.observe.choreo.gen.TelemetryOuterClass;
 import io.ballerina.observe.choreo.logging.LogFactory;
 import io.ballerina.observe.choreo.logging.Logger;
-import io.ballerina.observe.choreo.client.model.ChoreoMetric;
-import io.ballerina.observe.choreo.client.model.ChoreoTraceSpan;
-import io.ballerina.observe.choreo.client.model.SpanEvent;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import io.grpc.StatusRuntimeException;
